Create Parallel Switching Logic for Relays and Lights
I would like to credit sfromis for this code and the following description of how it works.

First line defines a class, and last line instantiates the class to create an object.
The function of init() is initialization code for when the class is instantiated (creating an object).
The tasmota.add_driver allows the class instance to receive callbacks from Tasmota in many different situations.
set_power_handler() is just one of those, and it is only to be notified when a power status change happens.
tasmota.get_power() accesses a list of power states for all channels
Index numbers in Berry starts with 0, and with 1 in Tasmota outside of Berry. Hence pow(2) matches Power3 etc.
tasmota.set_power() simply changes the state of one power channel.
Tricky thing is using the tasmota.set_timer(0, .... The reason is that the set_power_handler callback is not allowed to change any states, therefor the set_power call is deferred for 0 microseconds, meaning after exit from set_power_handler.

Berry code entered in the Berry REPL console is not saved anywhere. You can create a file autoexec.be with Berry code to be loaded after boot. When you use the console, the new code will in most cases supersede the old, but when using tasmota.add_driver the old iteration will persist. You can either reboot, build Tasmota with a development option providing a Berry restart button in the Berry REPL console, or you can change the init() method to:

  def init()
    tasmota.remove_driver(global.MasterPower_instance)
    global.MasterPower_instance = self
    tasmota.add_driver(global.MasterPower_instance)
  end

This allows the driver to remove the old version of itself, and then register the new instance.

The Backup Configuration function does not back up any files you have in the file system. So, you'll have to recreate the file for new devices, unless you do a custom build including a pre-populated file system.